chore: clippy 清理至零 warning
cargo clippy --fix 机械修复 11 文件(contains/iter().any()、冗余闭包、
多余引用、derive Default 等)+ 手动处理 3 处非机械项:
- session.rs: 抽 SessionExecContext type alias(消 type_complexity)
- fd.rs: close_unneeded_fds 改 enumerate().skip().take()(消 needless_range_loop)
- scheduler.rs: JobState::Done(JobResult) 保持内联 + #[allow] + rationale
  (数据在堆、并发有 max_concurrent 上限、避免终态无谓堆分配)

验证: cargo clippy --workspace --all-targets → 0 warning;
       cargo test --workspace → 313 passed / 0 failed

+++ crates/sandbox-core/src/process.rs
@@ -124,8 +124,8 @@ impl PreparedSandboxContext {
                     use std::os::unix::fs::MetadataExt;
                     if let Ok(meta) = workspace.metadata() {
                         let dev = meta.dev();
-                        io.major = (dev >> 8) as u64;
-                        io.minor = (dev & 0xff) as u64;
+                        io.major = dev >> 8;
+                        io.minor = dev & 0xff;
                     }
                 }
             }
